Align tauri import downloads with job queue
