+++ apps/tauri/src/infrastructure/api/clients/preset-client.ts
@@ -1,4 +1,4 @@
 import { createPresetClient } from "@solid-imager/ui/preset-client";
-import { orpc } from "./orpc-client";
+import { orpc } from "../../api-clients/orpc-client";
 
 export const PresetClient = createPresetClient(orpc);
